requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.3
djangorestframework==3.15.2
efficientnet-pytorch==0.7.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
numpy==2.1.3
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
pretrainedmodels==0.7.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
segmentation-models-pytorch==0.3.4
six==1.16.0
sqlparse==0.5.1
sympy==1.13.1
timm==0.9.7
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
